refactor: single reponsability to dashboard controller
